+++ R/export.r
@@ -1,3 +1,57 @@
+#' Explicitly declare module exports
+#'
+#' \code{box::export} can be used as an alternative to the \code{@export}
+#' tag comment to declare a module’s exports.
+#'
+#' @param ... zero or more unquoted names that should be exported from the
+#' module.
+#' @return \code{box::export} has no return value. It is called for its
+#' side-effect.
+#'
+#' @details
+#' \code{box::export} can be called inside a module to specify the module’s
+#' exports. If a module contains a call to \code{box::export}, this call
+#' overrides any declarations made via the \code{@export} tag comment. When a
+#' module contains multiple calls to \code{box::export}, the union of all thus
+#' defined names is exported.
+#'
+#' A module can also contain an argument-less call to \code{box::export}. This
+#' ensures that the module does not export any names. Otherwise, a module that
+#' defines names but does not mark them as exported would be treated as a
+#' \emph{legacy module}, and all default-visible names would be exported from
+#' it. Default-visible names are names not starting with a dot (\code{.}).
+#'
+#' @note The preferred way of declaring exports is via the \code{@export} tag
+#' comment. The main purpose of \code{box::export} is to prevent exports, by
+#' being called without arguments.
+#'
+#' @seealso
+#' \code{\link{use}} for information on declaring exports via \code{@export}.
+#' @export
+export = function (...) {
+    mod_ns = mod_topenv(parent.frame())
+    if (! is_namespace(mod_ns)) {
+        stop(sprintf('%s can only be called from inside a module', dQuote('export')))
+    }
+
+    names = as.list(match.call())[-1L]
+    for (name in names) {
+        if (! is.name(name)) {
+            stop(sprintf(
+                '%s requires a list of unquoted names; %s is not a name',
+                dQuote('export()'), dQuote(deparse(name))
+            ))
+        }
+    }
+
+    namespace_info(mod_ns, 'exports') = union(
+        namespace_info(mod_ns, 'exports'),
+        as.character(unlist(names))
+    )
+    namespace_info(mod_ns, 'legacy') = FALSE
+    invisible()
+}
+
 #' Find exported names in parsed module source
 #'
 #' @param info The module info.