+++ tests/testthat/mod/explicit_exports.r
@@ -1,0 +1,9 @@
+box::use(a = ./a[...])
+
+double = function (x) c(x, x)
+
+hidden = 'hidden'
+
+# module, local name, imported name
+#    └──────┐  └──┐       ┌─┘
+box::export(a, double, modname)